requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
boto3==1.21.8
botocore==1.24.8
click==8.0.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==0.10.0
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
redis==4.2.0
rq==1.10.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
wrapt==1.14.0
